Battery report changes for controller

+++ rmk/src/event.rs
@@ -150,7 +150,7 @@ pub enum ControllerEvent {
     /// Key event and action
     Key(KeyboardEvent, KeyAction),
     /// Battery percent changed
-    Battery(u16),
+    Battery(u8),
     /// Charging state changed
     ChargingState(bool),
     /// Ble profile changed
